backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==3.2.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
dj-rest-auth==5.0.2
Django==5.0.1
django-allauth==0.60.0
django-cors-headers==3.12.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
filelock==3.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
mysqlclient==2.1.0
oauthlib==3.2.2
openai==1.7.1
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.4.0
PyMySQL==1.1.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
websocket-client==1.7.0
zope.interface==6.1
